requirements.txt
requests==2.32.3
asgiref==3.8.1
Django==5.2
gunicorn==23.0.0
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
django-recaptcha==4.1.0
django-pg-copy==1.2.0
django-object-actions==5.0.0
easy_thumbnails==2.10
markdown==3.8.2
beautifulsoup4==4.13.4
django-mptt==0.17.0
django-multiupload==0.6.1
django-taggit==5.0.1
python-memcache==4.0.0
